# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Priya Global Platform API
  version: 1.0.0
  description: |
    # Priya Global - AI-Powered Multi-Channel Sales Platform

    Complete API documentation for the Priya Global Platform - a comprehensive multi-tenant system for managing customer communications, AI-powered chat, billing, analytics, and business operations across 36 microservices.

    ## Key Features
    - **Multi-tenant Architecture**: Fully isolated workspaces with granular access control
    - **36 Microservices**: Specialized services for auth, channels, AI, billing, analytics, and more
    - **Multi-channel Support**: WhatsApp, Email, Voice, SMS, Social, WebChat, Telegram, RCS, Video
    - **AI Integration**: Advanced conversational AI, sentiment analysis, and intelligent routing
    - **Enterprise Billing**: Stripe-powered subscriptions with regional pricing (USD, INR, GBP, AUD)
    - **Real-time Analytics**: Comprehensive dashboards and custom reporting
    - **Webhooks & Callbacks**: Event-driven architecture for real-time integrations
    - **Rate Limiting**: Plan-based throttling with per-tenant limits

    ## Authentication
    All endpoints use JWT (Bearer token) authentication except public webhooks.

    ```
    Authorization: Bearer {access_token}
    ```

    Access tokens are valid for 15 minutes. Use the refresh token to obtain new access tokens.

    ## Multi-tenancy
    Include tenant context in all requests:
...

    ## Rate Limiting
    Rates are plan-based:
...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      summary: User login
      tags: [Auth]
      description: Authenticate user and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token (15 minutes)
                  refresh_token:
                    type: string
                    description: Refresh token (7 days)
                  user_id:
                    type: string
                    format: uuid
                  workspace_id:
                    type: string
                    format: uuid
                  expires_in:
                    type: integer
                    description: Token expiration in seconds
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication. Obtain via /api/v1/auth/login
  schemas:
    Error:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        request_id:
          type: string
          format: uuid
          description: Unique request identifier for tracking
        timestamp:
          type: string
          format: date-time
...